--- app/main/page.tsx
@@ -1,5 +1,0 @@
-const Page = () => {
-  return <div>main</div>;
-};
-
-export default Page;
